# docs/swagger.yaml
basePath: /api
definitions:
  github_com_mkhsnw_golang-starter-kit_internal_model.CreateOrderRequest:
    properties:
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: integer
    required:
    - status
    - total_price
    - user_id
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.CreateProductRequest:
    properties:
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.LoginRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.OrderResponse:
    properties:
      created_at:
        type: integer
      id:
        type: integer
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: integer
      user_id:
        type: integer
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_item:
        type: integer
      total_page:
        type: integer
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.ProductResponse:
    properties:
      created_at:
        type: integer
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: integer
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.RegisterRequest:
    properties:
      email:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.TokenResponse:
    properties:
      token:
        type: string
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.UpdateOrderRequest:
    properties:
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: integer
    required:
    - status
    - total_price
    - user_id
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.UpdateProductRequest:
    properties:
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.UserResponse:
    properties:
      created_at:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: integer
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any:
    properties:
      data: {}
      errors:
        type: string
      paging:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata'
    type: object
  ? github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-array_github_com_mkhsnw_golang-starter-kit_internal_model_OrderResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.OrderResponse'
        type: array
      errors:
        type: string
      paging:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata'
    type: object
  ? github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-array_github_com_mkhsnw_golang-starter-kit_internal_model_ProductResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.ProductResponse'
        type: array
      errors:
        type: string
      paging:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata'
    type: object
  ? github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_OrderResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.OrderResponse'
      errors:
        type: string
      paging:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata'
    type: object
  ? github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_ProductResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.ProductResponse'
      errors:
        type: string
      paging:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata'
    type: object
  ? github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_TokenResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.TokenResponse'
      errors:
        type: string
      paging:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata'
    type: object
  github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_UserResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.UserResponse'
      errors:
        type: string
      paging:
        $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.PageMetadata'
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a sample API for Golang Starter Kit.
  title: Golang Starter Kit API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: Login Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_TokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      summary: Login user
      tags:
      - users
  /orders:
    get:
      description: Get a paginated list of orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-array_github_com_mkhsnw_golang-starter-kit_internal_model_OrderResponse'
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new Order
      parameters:
      - description: Create Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Create Order
      tags:
      - orders
  /orders/{id}:
    delete:
      description: Delete a Order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Delete Order
      tags:
      - orders
    get:
      description: Get a Order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_OrderResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Get Order by ID
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update an existing Order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Update Order
      tags:
      - orders
  /products:
    get:
      description: Get a paginated list of products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-array_github_com_mkhsnw_golang-starter-kit_internal_model_ProductResponse'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new Product
      parameters:
      - description: Create Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Create Product
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a Product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Delete Product
      tags:
      - products
    get:
      description: Get a Product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_ProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Get Product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing Product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Update Product
      tags:
      - products
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, and password
      parameters:
      - description: Register Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      summary: Register a new user
      tags:
      - users
  /users/current:
    get:
      consumes:
      - application/json
      description: Get current authenticated user details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-github_com_mkhsnw_golang-starter-kit_internal_model_UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mkhsnw_golang-starter-kit_internal_model.WebResponse-any'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"